backend/ai-service/requirements.txt
asgiref==3.8.1
Django==4.0.1
django-cors-headers==4.2.0
django-dotenv==1.4.2
djangorestframework==3.14.0
pillow==10.4.0
python-magic==0.4.27
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
